makefile

+++ noise.c
@@ -2,7 +2,6 @@
 
 int RangedRandDemo(int range_min, int range_max)
 {
-    // Generate random numbers in the interval [range_min, range_max], inclusive.
 
     int r = ((double)rand() / RAND_MAX) * (range_max - range_min) + range_min;
     return r;
@@ -12,6 +11,6 @@ void add_noise (double *radioactivity, int N)
 {
     for (int i = 0; i < N; i++)
     {
-        radioactivity[i] += (RangedRandDemo(-1000, 1000)/1000) * amplitude;
+        radioactivity[i] += ((double)RangedRandDemo(-1000, 1000)/1000) * amplitude;
     }
 }
